MainGame Init & CharacterStats, Battler Update
Created the MainGame Scence. From here the rest of the nodes will connect to form the basic structure of the game. Other key nodes include the BattleScene node and the WorldMap node.
Additionally, CharacterStats and Battlers scripts were updated. After consulting help online, methods for structuring the game and getting files to talk to each other is making more sense.
Note to self: be kind to self while learning how to build this multi-file system from scratch. Setting up the basic components of the game is necessary for data structures project, and tutorials are alright to 'set the stage' for the main focus of that project (specific data structs implemented). Keep it simple for now and create achievable baby-step goals to make it to the end - you can do it!