Character Stats Properly Allocated from Resource
I finally figured it out! The node path in the Battler Script needed to be corrected to be relative to the character scene: It was just looking for Stats and Skills in the tree, but because the Job scene which holds these was in the tree the compiler couldn't find it. So Job/Stats was what finally did it!
Now I don't have to do Direct Access for Stats and I can continue with Resource Access! Huzzah!