Turn Queue Progress
Data structure is built out, but is not executing properly. I think one solution is to build a Battle Manager Scene with would call the turn_tracker node as well as help to manage health bars and basic actions as well.
There is plenty more to implement before this project is complete, but I am making good progress.