UI and Attack Button Progress
Started working on attack functionality. Needed to sort out buttons and UI so that signals can be written in script easily.
So, UI is getting its own scene and build-out to be added to each scene. Buttons going literally anywhere at this point. - NOTE: If I were building this from scratch I would make much better plans for organizing code and layout of the game.
Coming up next is continuing on the attack functionality, possibly skill functionality, then managing the turns of each character in the battle. More to follow!